requirements.txt
python-telegram-bot==21.5
aiohttp==3.11.11
redis==5.2.1
flask==3.1.0
